+++ tests/testthat.R
@@ -2,5 +2,8 @@
 library(testthat)
 library(surveyvoi)
 
+# enable parallel testing
+Sys.unsetenv("R_TESTS")
+
 # run tests
 test_check("surveyvoi")